added universal CRUD class with static methods
universal CRUD class created with static methods for commonly used requests. for example CRUD.readAll will take the name of the collection and the req and res (generated in express API) connect to the database and return all from the named collection. 
This is so that code can be reused for later classes.